% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cattle_genome_table}
\alias{cattle_genome_table}
\title{Cattle genome table}
\format{
A data frame with three columns: the name of each chromosome,
the physical length in bp and the genetic length in centimorgans.
}
\source{
ARS-UCD1.2 genome assembly and Ma et al. (2015)
}
\usage{
cattle_genome_table
}
\description{
Physical and genetic length of cattle chromosomes
}
\keyword{datasets}
