% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{forward_simulate_population}
\alias{forward_simulate_population}
\title{Forward-simulate a population with a given sequence of population sizes.}
\usage{
forward_simulate_population(pop, pop_sizes, simparam)
}
\arguments{
\item{pop}{Starting population (AlphaSimR population object)}

\item{pop_sizes}{Vector of population sizes for each generation}

\item{simparam}{Simulation parameters (AlphaSimR SimParam object)}
}
\value{
A list of populations, representing each generation.
}
\description{
Forward-simulate a population with a given sequence of population sizes.
}
