% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_history.R
\name{generations_ago_to_pop_size}
\alias{generations_ago_to_pop_size}
\title{Turn a data frame of population sizes generations at generations ago into
a vector of population sizes for forward simulation.}
\usage{
generations_ago_to_pop_size(history)
}
\arguments{
\item{history}{A data frame with columns Ne and generations_ago.}
}
\value{
A vector of population sizes.
}
\description{
Turn a data frame of population sizes generations at generations ago into
a vector of population sizes for forward simulation.
}
