% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_map.R
\name{make_adjusted_map}
\alias{make_adjusted_map}
\title{Adjust a simulated linkage map by interpolating positions on a real
linkage map}
\usage{
make_adjusted_map(sim_map, real_map)
}
\arguments{
\item{sim_map}{List of maps for each chromosome in Morgans}

\item{real_map}{List of maps for each chromosome in Morgans}
}
\value{
A list containing the modified map
}
\description{
Adjust a simulated linkage map by interpolating positions on a real
linkage map
}
