% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pig_genome_table}
\alias{pig_genome_table}
\title{Pig genome table}
\format{
A data frame with three columns: the name of each chromosome,
the physical length in bp and the genetic length in centimorgans.
}
\source{
Sscrofa11.1 genome assembly and Tortereau et al. (2012)
}
\usage{
pig_genome_table
}
\description{
Physical and genetic length of cattle chromosomes
}
\keyword{datasets}
