% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_founder_genome}
\alias{simulate_founder_genome}
\title{Simulate a founder genome based on a genome table.}
\usage{
simulate_founder_genome(
  genome_table,
  n_ind,
  seg_sites_per_bp,
  final_Ne,
  historical_Ne,
  historical_Ne_time,
  split,
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{genome_table}{Data frame of genome parameters, must contain columns
"length" with chromosome length in bp and "genetic lengths" with genetic
length in cM}

\item{n_ind}{Number of individual}

\item{seg_sites_per_bp}{Optional, number of segregating sites to be kept per
basepair. NULL for all}

\item{final_Ne}{Final effective population size. Passed to AlphaSimR.}

\item{historical_Ne}{Optional historical population sizes. Passed to AlphaSimR.}

\item{historical_Ne_time}{Optional timings for population size changes.
Passed to AlphaSimR.}

\item{split}{Optional split parameter. Passed to AlphaSimR.}
}
\value{
An AlphaSimR PopMap object to be used as founder of a simulation
}
\description{
Simulate a founder genome based on a genome table.
}
